Move AssetTracker to new namespace. Resolve orchestrator startup issue
